requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
fake-useragent==2.0.2
Flask==3.1.0
Flask-RESTful==0.3.10
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
packaging==24.2
parse==1.20.2
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-html==0.10.0
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
w3lib==2.2.1
websockets==10.4
Werkzeug==3.1.3
zipp==3.21.0
